+++ cmd/alter/alter.go
@@ -9,10 +9,11 @@ func NewAlterCmd() *cobra.Command {
 	var cmdAlter = &cobra.Command{
 		Use:     "alter",
 		Aliases: []string{"edit"},
-		Short:   "alter topics, partitions",
+		Short:   "alter topics, partitions, brokers",
 	}
 
 	cmdAlter.AddCommand(newAlterTopicCmd())
 	cmdAlter.AddCommand(newAlterPartitionCmd())
+	cmdAlter.AddCommand(newAlterBrokerCmd())
 	return cmdAlter
 }